feat: Enhance `ReporteProductividad` and `ReporteLecturas` components, add new SQL scripts for daily and by-date readings, and update build artifacts and dependencies

+++ DieselApp/src/components/ReporteLecturas.tsx
@@ -562,7 +562,7 @@ export default function ReporteLecturas() {
                               onChange={(e) =>
                                 setEditForm({
                                   ...editForm,
-                                  LecturaCms: parseInt(e.target.value) || 0,
+                                  LecturaCms: parseFloat(e.target.value) || 0,
                                 })
                               }
                             />
